refactor(web): extract filter primitives for reuse
Move MultiSelect, YearRange, SearchInput, ResetButton out of the map's
FilterBar into web/src/components/filters/primitives.tsx so they can be
reused on the companies directory page (and any future filter surface).

The map's FilterBar now imports these primitives and only owns its own
positioning + count badge. No behavior change.
